feat: added extensive testcase for agentkit
