+++ api/assistant-api/internal/llm/agentkit/type.go
@@ -8,7 +8,10 @@ package internal_llm_agentkit
 
 import (
 	internal_type "github.com/rapidaai/api/assistant-api/internal/type"
+	"github.com/rapidaai/pkg/commons"
 	"github.com/rapidaai/protos"
+	"google.golang.org/grpc"
+	"sync"
 )
 
 type AgentPipeline interface {
@@ -33,6 +36,20 @@ type InterruptionPipeline struct {
 
 type ResponsePipeline struct {
 	Response *protos.TalkOutput
+}
+
+type agentkitExecutor struct {
+	logger          commons.Logger
+	transport       agentkitTransport
+	stateMu         sync.RWMutex
+	activeContextID string
+	closing         bool
+}
+
+type agentkitTransport struct {
+	conn         *grpc.ClientConn
+	stream       grpc.BidiStreamingClient[protos.TalkInput, protos.TalkOutput]
+	listenerDone chan struct{}
 }
 
 func (UserTurnPipeline) agentPipeline()      {}
